PDF to docx converter
This project was done by me when I was a Intern at Kidocode on September 3rd week of 2024.

This code usually converts the PDF file into a doc format without any overlapping or wrapping text inside.

This code uses the terminal or you may say the console application . It cannot comply with images and other attrbutes inside word.